Force AUFS storage driver instead of devicemapper
In order to workaround docker/docker#4036 which prevents us from using
`tsuru-admin platform-add` without multiple attempts:

     ---> d26617e2bb8b
    Removing intermediate container 3ec79f2117be
    Step 8 : RUN mkdir /var/lib/tsuru
     ---> Running in fedd00a99921
    error in docker node "http://10.128.1.190:4243": Error getting container fedd00a999218247ec4721d14d5cc446b172c861eb39d58ba1468d408a7b3c19 from driver devicemapper: open /dev/mapper/docker-202:1-524978-fedd00a999218247ec4721d14d5cc446b172c861eb39d58ba1468d408a7b3c19: no such file or directory
    Error: Failed to add new platform.

On machines which don't have the necessary kernel modules for AUFS support
(such as Ubuntu 14.04 AMIs on AWS which don't have the kernel extras
package) this will cause Docker to fail on start with the error:

    Shutting down daemon due to errors: error intializing graphdriver: driver not supported

This happens too late for upstart to notice, so Ansible won't
register the failure and will attempt to start it again on the next run.
However it will render Docker unusable so we should notice that the right
driver isn't available.