Install kernel extra package for AUFS support
So that we're able to use the right storage driver per the previous commit.
This mimics docker/docker@b89d04dd62e1a779f449804772c5bc5b64af65e8 and the
issue that references, which explains why we need to install both packages.

It does not require a machine reboot to take effect but it does need to
restart the docker daemon.