Add default socket to listen options
Reinstate the default UNIX socket which Docker listens on if you don't
specify different listen options. We disabled this when opening up the TCP
port for Tsuru to talk use.

This meant that `sudo docker <command>` no longer worked unless you export a
`DOCKER_HOST` environment variable or pass a `-H` argument to each command,
which was a pain.