Docker using virtual packages for latest version
Docker uses a meta package delivery method to encourage you to use the
latest version of docker. What they do is name older packages by
appending the version name to it with a hyphen rather than a version.

To work around this I've refactored the code to make it handle the hyphen
and be more extensible so you can mix and match it when using
`with_items` e.g.

- name: Install support packages
  apt: name="{{ item }}"
  with_items:
    - foo
    - "{% if bar_version is defined %}bar={{ bar_version }}{% else %}bar{% endif %}
    - "{% if qux_version is defined %}qux-{{ qux_version }}{% else %}qux{% endif %}